Reorganize files in project and make some code style fixes.
